refactor: 변수 이름 의미있게 변경 & 함수 분리 & 컴포넌트 분리
